requirements.txt
# ================================================================
# PRODUCTION REQUIREMENTS FOR CONVERTER APPLICATION
# ================================================================
# This file contains all dependencies needed to run the converter
# application in production. For development dependencies, see
# requirements-dev.txt

# === CORE FLASK FRAMEWORK ===
Flask==3.0.0
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
ItsDangerous==2.1.2
Click==8.1.7

# === WEB SERVER AND DEPLOYMENT ===
gunicorn==21.2.0
brotli==1.1.0

# === DATABASE AND ORM ===
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
Flask-Migrate==4.0.5
Alembic==1.13.0
psycopg2-binary==2.9.9

# === AUTHENTICATION AND SECURITY ===
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
bcrypt==4.1.2
argon2-cffi==23.1.0
cryptography==41.0.8
PyJWT==2.8.0

# Rate limiting and session management
Flask-Limiter==3.5.0
Flask-Session==0.5.0
redis==5.0.1
hiredis==2.2.3

# CSRF protection
Flask-SeaSurf==1.1.1

# === ENVIRONMENT CONFIGURATION ===
python-dotenv==1.0.0

# === FILE PROCESSING CORE ===
# Image processing
Pillow==10.1.0
numpy==1.26.2

# PDF processing
PyMuPDF==1.23.8
pdf2docx==0.5.6
PyPDF2==3.0.1
pypdf==3.17.4
pdfplumber==0.10.3
pdf2image==1.16.3

# Document processing
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4

# === SPECIALIZED CONVERTERS ===
# RAW image processing
rawpy==0.25.1

# HEIC image support
pillow-heif==0.13.1

# Video and GIF processing
imageio==2.33.1
imageio-ffmpeg==0.4.9
opencv-python==4.8.1.78

# Configuration file processing
PyYAML==6.0.1
ruamel.yaml==0.18.5

# HTML to PDF conversion
weasyprint==60.2
pdfkit==1.0.0
selenium==4.16.0
webdriver-manager==4.0.1

# Jupyter Notebook processing
nbformat==5.9.2
nbconvert==7.16.4
jupyter-client==8.6.0
jupyter-core==5.5.0
traitlets==5.14.0
mistune==3.0.2
pygments==2.17.2
bleach==6.1.0
defusedxml==0.7.1
entrypoints==0.4
packaging==23.2

# Pandoc integration
pypandoc==1.13
pypandoc-binary==1.13

# === FILE TYPE DETECTION AND VALIDATION ===
python-magic==0.4.27
filetype==1.2.0
chardet==5.2.0

# === WEB AND MARKUP PROCESSING ===
markdown==3.5.1
markdown2==2.4.12
html2text==2020.1.16
beautifulsoup4==4.12.2

# === OCR AND TEXT EXTRACTION ===
pytesseract==0.3.10

# === ADVANCED DOCUMENT PROCESSING ===
reportlab==4.0.7
xlsxwriter==3.1.9
xlrd==2.0.1
python-pptx==0.6.23

# === QR CODE GENERATION ===
qrcode[pil]==7.4.2

# === TASK SCHEDULING ===
APScheduler==3.10.4

# === HTTP CLIENT AND NETWORKING ===
requests==2.31.0
urllib3==2.1.0

# === EMAIL SERVICES ===
Flask-Mail==0.9.1
email-validator==2.1.0
DNSpython==2.4.2

# === MONITORING AND LOGGING ===
sentry-sdk[flask]==1.39.2
prometheus-flask-exporter==0.23.0
Flask-HealthZ==1.0.1
Flask-UUID==0.2

# === VALIDATION AND SERIALIZATION ===
marshmallow==3.20.2
webargs==8.4.0

# === BACKGROUND TASK PROCESSING ===
celery==5.3.4

# === TWO-FACTOR AUTHENTICATION ===
pyotp==2.9.0

# === SCIENTIFIC COMPUTING ===
scipy==1.11.4

# === CLOUD STORAGE (Optional) ===
# Uncomment as needed for cloud deployment
# boto3==1.34.0              # AWS S3
# azure-storage-blob==12.19.0 # Azure Blob Storage
# google-cloud-storage==2.13.0 # Google Cloud Storage

# === SYSTEM DEPENDENCIES NOTES ===
# The following require system-level installation:
# - FFmpeg (for video processing)
# - Tesseract OCR (for text extraction)
# - wkhtmltopdf (for HTML to PDF conversion)
# - Pandoc (for document conversion)
# - LaTeX distribution (for LaTeX to PDF conversion)
#
# Install commands (adjust for your OS):
# Windows: 
#   choco install ffmpeg tesseract wkhtmltopdf pandoc miktex
# macOS:
#   brew install ffmpeg tesseract-ocr wkhtmltopdf pandoc mactex
# Ubuntu/Debian:
#   sudo apt-get install ffmpeg tesseract-ocr wkhtmltopdf pandoc texlive-latex-recommended
